[Operator Versioning] Reorganize unittest
Differential Revision: [D33180767](https://our.internmc.facebook.com/intern/diff/D33180767/)

[ghstack-poisoned]
